+++ ui/src/containers/VisualiseForm/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import StatementsForm from './StatementsForm/StatementsForm';
+import StatementsForm from './StatementsForm';
 
 const VisualiseForm = ({ model, exportVisualisation }) =>
   <StatementsForm model={model} exportVisualisation={exportVisualisation} />;
